requirements.txt
Django>=1.8,<2.0
psycopg2
djangorestframework
django-filter
django-extensions
django-admin-sortable2
django-cors-headers
mixer
ipdb
requests
